requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.28
pandas==2.1.3
matplotlib==3.8.2
requests==2.31.0
scikit-learn==1.3.2
pymongo==4.6.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
plotly==5.17.0
python-dotenv==1.0.0